+++ BookJam/Source/DetailPage/Model/MainDetailTableSection.swift
@@ -1,0 +1,16 @@
+//
+//  MainDetailTableSection.swift
+//  BookJam
+//
+//  Created by 장준모 on 12/27/23.
+//
+
+import Foundation
+
+enum SettingSection {
+    case home
+    case news(title: String)
+    case activity
+    case review
+    case bookList
+}
